+++ packages/iov-cosmoshub/src/encode.spec.ts
@@ -11,7 +11,7 @@ import {
   SignedTransaction,
   TokenTicker,
 } from "@iov/bcp";
-import { Encoding } from "@iov/encoding";
+import { fromBase64 } from "@iov/encoding";
 
 import {
   buildSignedTx,
@@ -21,8 +21,6 @@ import {
   encodeFullSignature,
   encodePubkey,
 } from "./encode";
-
-const { fromBase64 } = Encoding;
 
 describe("encode", () => {
   const atom = "ATOM" as TokenTicker;