+++ packages/iov-encoding/types/index.d.ts
@@ -1,3 +1,4 @@
+export { fromBase64, toBase64 } from "./base64";
 export { Bech32 } from "./bech32";
 export { Decimal } from "./decimal";
 export { Encoding } from "./encoding";