+++ packages/iov-tendermint-rpc/src/v0-32/hasher.spec.ts
@@ -1,11 +1,9 @@
-import { Encoding, fromHex } from "@iov/encoding";
+import { fromBase64, fromHex } from "@iov/encoding";
 import { ReadonlyDate } from "readonly-date";
 
 import { ReadonlyDateWithNanoseconds } from "../responses";
 import { TxBytes } from "../types";
 import { hashBlock, hashTx } from "./hasher";
-
-const { fromBase64 } = Encoding;
 
 describe("Hasher", () => {
   it("creates transaction hash equal to local test", () => {